+++ src/Symfony/Component/EventDispatcher/Tests/DependencyInjection/RegisterListenersPassTest.php
@@ -13,12 +13,17 @@ namespace Symfony\Component\EventDispatcher\Tests\DependencyInjection;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\Argument\ServiceClosureArgument;
+use Symfony\Component\DependencyInjection\Compiler\AttributeAutoconfigurationPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\EventDispatcher\DependencyInjection\AddEventAliasesPass;
 use Symfony\Component\EventDispatcher\DependencyInjection\RegisterListenersPass;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\EventDispatcher\Tests\Fixtures\CustomEvent;
+use Symfony\Component\EventDispatcher\Tests\Fixtures\TaggedInvokableListener;
+use Symfony\Component\EventDispatcher\Tests\Fixtures\TaggedMultiListener;
+use Symfony\Contracts\Service\Attribute\TagInterface;
 
 class RegisterListenersPassTest extends TestCase
 {
@@ -231,6 +236,82 @@ class RegisterListenersPassTest extends TestCase
         $this->assertEquals($expectedCalls, $definition->getMethodCalls());
     }
 
+    /**
+     * @requires PHP 8
+     */
+    public function testTaggedInvokableEventListener()
+    {
+        if (!\interface_exists(TagInterface::class)) {
+            self::markTestSkipped('This test requires symfony/service-contracts 2.4 or newer.');
+        }
+
+        $container = new ContainerBuilder();
+        $container->register('foo', TaggedInvokableListener::class)->setAutoconfigured(true);
+        $container->register('event_dispatcher', \stdClass::class);
+
+        (new AttributeAutoconfigurationPass())->process($container);
+        (new RegisterListenersPass())->process($container);
+
+        $definition = $container->getDefinition('event_dispatcher');
+        $expectedCalls = [
+            [
+                'addListener',
+                [
+                    CustomEvent::class,
+                    [new ServiceClosureArgument(new Reference('foo')), '__invoke'],
+                    0,
+                ],
+            ],
+        ];
+        $this->assertEquals($expectedCalls, $definition->getMethodCalls());
+    }
+
+    /**
+     * @requires PHP 8
+     */
+    public function testTaggedMultiEventListener()
+    {
+        if (!\interface_exists(TagInterface::class)) {
+            self::markTestSkipped('This test requires symfony/service-contracts 2.4 or newer.');
+        }
+
+        $container = new ContainerBuilder();
+        $container->register('foo', TaggedMultiListener::class)->setAutoconfigured(true);
+        $container->register('event_dispatcher', \stdClass::class);
+
+        (new AttributeAutoconfigurationPass())->process($container);
+        (new RegisterListenersPass())->process($container);
+
+        $definition = $container->getDefinition('event_dispatcher');
+        $expectedCalls = [
+            [
+                'addListener',
+                [
+                    CustomEvent::class,
+                    [new ServiceClosureArgument(new Reference('foo')), 'onCustomEvent'],
+                    0,
+                ],
+            ],
+            [
+                'addListener',
+                [
+                    'foo',
+                    [new ServiceClosureArgument(new Reference('foo')), 'onFoo'],
+                    42,
+                ],
+            ],
+            [
+                'addListener',
+                [
+                    'bar',
+                    [new ServiceClosureArgument(new Reference('foo')), 'onBarEvent'],
+                    0,
+                ],
+            ],
+        ];
+        $this->assertEquals($expectedCalls, $definition->getMethodCalls());
+    }
+
     public function testAliasedEventListener()
     {
         $container = new ContainerBuilder();
@@ -413,10 +494,6 @@ final class AliasedSubscriber implements EventSubscriberInterface
 }
 
 final class AliasedEvent
-{
-}
-
-final class CustomEvent
 {
 }