+++ src/Symfony/Component/HttpKernel/Tests/Fixtures/ClearableService.php
@@ -2,6 +2,9 @@
 
 namespace Symfony\Component\HttpKernel\Tests\Fixtures;
 
+use Symfony\Component\HttpKernel\Attribute\Reset;
+
+#[Reset(method: 'clear')]
 class ClearableService
 {
     public static $counter = 0;