+++ src/Symfony/Component/HttpKernel/Tests/Fixtures/ResettableService.php
@@ -2,6 +2,9 @@
 
 namespace Symfony\Component\HttpKernel\Tests\Fixtures;
 
+use Symfony\Component\HttpKernel\Attribute\Reset;
+
+#[Reset(method: 'reset')]
 class ResettableService
 {
     public static $counter = 0;